apps/telegram-bot/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finsight-telegram-bot"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "python-telegram-bot[ext]>=20.7",
  "httpx>=0.27",
  "structlog>=24.1",
  "celery[redis]>=5.3",
  "pydantic>=2.7",
  "pyyaml>=6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
